requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
PyJWT>=2.8
